backend/requirements.txt
django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
pytest==8.3.4
pytest-django==4.9.0
